refactor: extract BufferState + initial_delivery_mode to buffer_state.rs (#62)
endpoint_task.rs reached 1013 lines (> 1000 limit). Extracted BufferState,
initial_delivery_mode, and their tests to a new buffer_state module with
focused responsibility: shared state types and pure decision helpers.

endpoint_task.rs now 981 lines. Test files still see BufferState via
pub use re-export from endpoint_task.
